Add files via upload
This is the idea of schematics behind DzigaLoop. It can be realized in different ways, even if I suggest to use double sided PCB for two hubs, one for 5V cables and one for GND cables. Then you can use just one cable to connect each hub to Bela Mini, and in this way you don't have to stretch to many cables and the lack of holes on the board is solved.